db: move excise tests
